+++ includes/Utils.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace MediaWiki\Extension\AchievementBadges;
+
+use BetaFeatures;
+use MediaWiki\MediaWikiServices;
+use User;
+
+class Utils {
+	/**
+	 * @param User $user
+	 * @return bool
+	 */
+	public static function isAchievementBadgesAvailable( User $user ) {
+		$config = MediaWikiServices::getInstance()->getMainConfig();
+
+		$configEnabled = $config->get( Constants::CONFIG_KEY_ACHIEVEMENT_BADGES_ENABLE_BETA_FEATURE );
+		$userOptionEnabled = $configEnabled &&
+			BetaFeatures::isFeatureEnabled( $user, Constants::PREF_KEY_ACHIEVEMENT_ENABLE );
+
+		if ( !$configEnabled ) {
+			// If AchievementBadges is not a beta feature, it is available to everyone.
+			return true;
+		}
+		if ( $user->isRegistered() && $userOptionEnabled ) {
+			// If AchievementBadges is a beta feature, only a registered user which enables the feature
+			// has access the feature.
+			return true;
+		}
+		return false;
+	}
+}